Tests for the genesis time config option for truffle test.
