+++ packages/truffle/test/sources/genesis_time/genesis_time_invalid/truffle-config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  genesis_time: "1/33/18"
+};
